+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# HEAD
+
+- Changed: no longer required to transfer options for `modernizr` using `config` option.
+- Changed: to load configuration from file, need specify `useConfigFile` option.
+- Chore: no need to use function `encodeURI` for `query string`.
+
 # 1.0.2 - 2017-01-31
 
 - Fixed: updated webpack peer dependency to support `2.2.1`.